+++ {{cookiecutter.repo_name}}/project/mods_available/staticpages/settings.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 INSTALLED_APPS = add_to_tuple(INSTALLED_APPS, 'staticpages')
 
+ENABLE_STATICPAGES = True
+
 STATICPAGES_PROTOTYPES = (
     (r'prototypes/$', "prototypes/index.html", 'prototypes-index'),
     (r'prototypes/buttons/$', "prototypes/buttons.html", 'prototypes-buttons'),